LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/intrinsic-iteration-duration.tentative-expected.txt


PASS Computed duration in percent even when specified in ms
FAIL Time-based duration normalized to fill animation range. assert_approx_equals: values do not match for "undefined" expected 60 +/- 0.125 but got 100
PASS Time-based duration normalized to preserve proportional delays.
PASS Intrinsic iteration duration fills timeline.
FAIL Intrinsic iteration duration accounts for animation range. assert_approx_equals: values do not match for "undefined" expected 60 +/- 0.125 but got 100
PASS Intrinsic iteration duration accounts for number of iterations
